requirements.txt

Flask
google-cloud-texttospeech
moviepy
